ixww22: Change ccuid in EEPROM to 8 bytes
Due to wrong calculation, we need to extend the CCUID length to 8 bytes
for full QR code support.

Testing and Verification Instructions:
1. set UNIT_TEST_EEPROM to 1
2. set TEST_EEPM_DATA_ID to TEST_EEPM_DATA_QRCODE_CCUID
3. check if console there's write and read success

Type: Bug Fix
Issue: AUTOOEM-IXWW22-6050
Reviewed-on: https://gerrit.garmin.com.tw/code-review/c/autooem/products/nxp/rt1172/+/85658
Approve-Change: Liao, Edward <edward.liao@garmin.com>
Tested-by: Huang, Kenny <kenny.huang@garmin.com>
Tested-by: Liao, Edward <edward.liao@garmin.com>

+++ eeprom/eep_prv.h
@@ -175,13 +175,13 @@ BaseType_t eep_get_burn_in_target_time
 
 BaseType_t eep_set_ccu_id
     (
-    uint32_t* ccu_id_ptr,
+    uint8_t* ccu_id_ptr,
     void ( *callback_func_ptr ) ( status_t )
     );
 
 BaseType_t eep_get_ccu_id
     (
-    uint32_t* ccu_id_ptr,
+    uint8_t* ccu_id_ptr,
     void ( *callback_func_ptr ) ( status_t )
     );
